requirements.txt
Django==1.11
gunicorn==19.6.0
pillow==3.4.2
